requirements.txt
requests>=2.31.0
colorama>=0.4.6
prompt_toolkit>=3.0.36
pyyaml>=6.0
click>=8.1.0
pyperclip>=1.8.2